rename components, step up final step page
