+++ src/components/DeliveryOptions/Email.jsx
@@ -13,10 +13,6 @@ function Email(props) {
         <p>Sign up here for reminders. By entering your phone number, you consent to receiving messages from vote.gov.</p>
         <Label htmlFor="sms-alert-phone-number">Opt in for SMS Alerts (optional)</Label>
         <TextInput id="sms-alert-phone-number" name="sms-alert-phone-number" value={props.fieldData.sms_alert_phone_number} onChange={props.saveFieldData('sms_alert_phone_number')} type="text" autoComplete="off" required={false}/>
-
-        <Button type="submit" onClick={props.handleNext}>
-            Send Form to My Email Address
-        </Button>
         </>
     );
 }
