+++ src/components/DeliveryOptions/Print.jsx
@@ -10,10 +10,6 @@ function Print(props) {
         <p>Sign up here for reminders. By entering your phone number, you consent to receiving messages from vote.gov.</p>
         <Label htmlFor="sms-alert-phone-number">Opt in for SMS Alerts (optional)</Label>
         <TextInput id="sms-alert-phone-number" name="sms-alert-phone-number" value={props.fieldData.sms_alert_phone_number} onChange={props.saveFieldData('sms_alert_phone_number')} type="text" autoComplete="off" required={false}/>
-
-        <Button type="submit" onClick={props.handleNext}>
-            Print Form
-        </Button>
         </>
     );
 }
